lecture_10/requirements.txt
flask
flask-restplus
flask-sqlalchemy
flask-marshmallow
marshmallow